requirements.txt
asgiref==3.5.2
Django==4.0.4
django-autoslug==1.9.8
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
